backend/requirements.txt

Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.4
reportlab==4.0.7
openpyxl==3.1.2
djangorestframework-simplejwt==5.3.1